Resolve 0U ambiguity in string utility tests

+++ include/etl/string_utilities.h
@@ -745,7 +745,7 @@ namespace etl
     if (required_size > s.size())
     {
       required_size -= s.size();
-      s.insert(0U, required_size, pad_char);
+      s.insert(size_t(0U), required_size, pad_char);
     }
   }
