+++ test/test_string_utilities_u16.cpp
@@ -1086,7 +1086,7 @@ namespace
     {
       String text(STR("Hello World"));
       String expected(text);
-      expected.insert(0U, text.capacity() - expected.size(), STR('x'));
+      expected.insert(size_t(0U), text.capacity() - expected.size(), STR('x'));
 
       etl::pad_left(text, text.capacity() + 1U, STR('x'));
 
@@ -1154,7 +1154,7 @@ namespace
     {
       String text(STR("Hello World"));
       String expected(text);
-      expected.insert(0U, text.capacity() - expected.size(), STR('x'));
+      expected.insert(size_t(0U), text.capacity() - expected.size(), STR('x'));
 
       etl::pad(text, text.capacity() + 1U, etl::string_pad_direction::LEFT, STR('x'));
